requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
psycopg==3.1.18
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
